removed useless type conversion (#17)
* removed useless type conversion

* removed useless type conversion

* byte[]

+++ printer.go
@@ -27,7 +27,7 @@ func AppendFixedPointDecimal(b []byte, v int64, p int) []byte {
 	}
 
 	if p == 0 {
-		return strconv.AppendInt(b, int64(v), 10)
+		return strconv.AppendInt(b, v, 10)
 	}
 
 	if v < 0 {
@@ -36,7 +36,7 @@ func AppendFixedPointDecimal(b []byte, v int64, p int) []byte {
 	}
 
 	s := len(b)
-	b = strconv.AppendInt(b, int64(v), 10)
+	b = strconv.AppendInt(b, v, 10)
 
 	if len(b)-s > p {
 		i := len(b) - p
